apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.39.9",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-jazzmin>=3.0.1",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
]
